Rename http server Lua module and fix its documentation (#2594)
